De-duplicate WizardTaskItem
