+++ client/vat/form/tasks/vat-number-task.tsx
@@ -160,6 +160,7 @@ export const VatNumberTask = ( {
 				getVatTaxIDName()
 			) }
 			className={ null }
+			visibleDescription={ null }
 		>
 			<p className="wcpay-wizard-task__description-element">
 				{ __(
