--- multi-currency/client/setup/wizard/task-item.scss
@@ -1,17 +1,0 @@
-.wcpay-wizard-task {
-	&__visible-description-element {
-		position: absolute;
-		margin-left: 40px;
-		margin-top: 0;
-		margin-bottom: 1em;
-
-		&.is-muted-color {
-			color: $gray-700;
-		}
-
-		.components-external-link svg {
-			width: 1em;
-			height: 1em;
-		}
-	}
-}
